Remove SchemaStore trait
