+++ aquadoggo/src/db/mod.rs
@@ -9,7 +9,6 @@ pub mod errors;
 pub mod models;
 pub mod provider;
 pub mod stores;
-pub mod traits;
 pub mod utils;
 
 /// Re-export of generic connection pool type.
