+++ aquadoggo/src/db/stores/schema.rs
@@ -2,17 +2,16 @@
 
 use std::convert::{TryFrom, TryInto};
 
-use async_trait::async_trait;
 use p2panda_rs::document::DocumentViewId;
 use p2panda_rs::schema::system::{SchemaFieldView, SchemaView};
 use p2panda_rs::schema::{Schema, SchemaId};
+use p2panda_rs::storage_provider::error::OperationStorageError;
+use sqlx::query_scalar;
 
 use crate::db::errors::SchemaStoreError;
 use crate::db::provider::SqlStorage;
-use crate::db::traits::SchemaStore;
 
-#[async_trait]
-impl SchemaStore for SqlStorage {
+impl SqlStorage {
     /// Get a Schema from the database by it's document view id.
     ///
     /// Internally, this method performs three steps:
@@ -22,7 +21,7 @@ impl SchemaStore for SqlStorage {
     ///
     /// If no schema definition with the passed id is found then None is returned, if any of the
     /// other steps can't be completed, then an error is returned.
-    async fn get_schema_by_id(
+    pub async fn get_schema_by_id(
         &self,
         id: &DocumentViewId,
     ) -> Result<Option<Schema>, SchemaStoreError> {
@@ -58,7 +57,7 @@ impl SchemaStore for SqlStorage {
     /// Returns an error if a fatal db error occured.
     ///
     /// Silently ignores incomplete or broken schema definitions.
-    async fn get_all_schema(&self) -> Result<Vec<Schema>, SchemaStoreError> {
+    pub async fn get_all_schema(&self) -> Result<Vec<Schema>, SchemaStoreError> {
         let schema_views: Vec<SchemaView> = self
             .get_latest_document_views_by_schema(&SchemaId::new("schema_definition_v1")?)
             .await?
@@ -88,10 +87,37 @@ impl SchemaStore for SqlStorage {
 
         Ok(all_schema.into_iter().flatten().collect())
     }
+
+    /// Returns the schema id for a document view.
+    ///
+    /// Returns `None` if this document view is not found.
+    pub async fn get_schema_by_document_view(
+        &self,
+        view_id: &DocumentViewId,
+    ) -> Result<Option<SchemaId>, SchemaStoreError> {
+        let result: Option<String> = query_scalar(
+            "
+            SELECT
+                schema_id
+            FROM
+                document_views
+            WHERE
+                document_view_id = $1
+            ",
+        )
+        .bind(view_id.to_string())
+        .fetch_optional(&self.pool)
+        .await
+        .map_err(|e| OperationStorageError::FatalStorageError(e.to_string()))?;
+
+        // Unwrap because we expect no invalid schema ids in the db.
+        Ok(result.map(|id_str| id_str.parse().unwrap()))
+    }
 }
 
 #[cfg(test)]
 mod tests {
+    use p2panda_rs::document::DocumentViewId;
     use p2panda_rs::identity::KeyPair;
     use p2panda_rs::schema::{FieldType, SchemaId};
     use p2panda_rs::test_utils::fixtures::{key_pair, random_document_view_id};
@@ -100,8 +126,6 @@ mod tests {
     use crate::db::stores::test_utils::{
         add_document, add_schema, test_db, TestDatabase, TestDatabaseRunner,
     };
-
-    use super::SchemaStore;
 
     #[rstest]
     fn get_schema(key_pair: KeyPair, #[from(test_db)] runner: TestDatabaseRunner) {
@@ -176,6 +200,56 @@ mod tests {
 
             // We receive nothing as the fields are missing for this schema
             assert!(schema.is_none());
+        });
+    }
+
+    #[rstest]
+    fn test_get_schema_for_view(
+        key_pair: KeyPair,
+        #[from(test_db)]
+        #[with(1, 1, 1)]
+        runner: TestDatabaseRunner,
+    ) {
+        runner.with_db_teardown(|mut db: TestDatabase| async move {
+            let schema = add_schema(
+                &mut db,
+                "venue",
+                vec![
+                    ("description", FieldType::String),
+                    ("profile_name", FieldType::String),
+                ],
+                &key_pair,
+            )
+            .await;
+
+            let document_view_id = match schema.id() {
+                SchemaId::Application(_, view_id) => view_id,
+                _ => panic!("Invalid schema id"),
+            };
+
+            let result = db.store.get_schema_by_document_view(document_view_id).await;
+
+            assert!(result.is_ok());
+            // This is the schema name of the schema document we published.
+            assert_eq!(result.unwrap().unwrap().name(), "schema_definition");
+        });
+    }
+
+    #[rstest]
+    fn test_get_schema_for_missing_view(
+        random_document_view_id: DocumentViewId,
+        #[from(test_db)]
+        #[with(1, 1, 1)]
+        runner: TestDatabaseRunner,
+    ) {
+        runner.with_db_teardown(|db: TestDatabase| async move {
+            let result = db
+                .store
+                .get_schema_by_document_view(&random_document_view_id)
+                .await;
+
+            assert!(result.is_ok());
+            assert!(result.unwrap().is_none());
         });
     }
 }
