--- aquadoggo/src/db/traits/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod schema;
-
-pub use schema::SchemaStore;
